requirements.txt
bottle==0.12.23
dalybms==0.3.0
pyserial==3.5
python-dotenv==0.21.0
